Deleted "The Version 1" classes of REST, with converters. REST service for Event only works correctly locally. UserResource, ProductResource and OrderResource should be fixed.
