+++ src/main/java/ru/geekbrains/spring/ishop/config/AppConfig.java
@@ -4,14 +4,12 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
-import org.springframework.format.FormatterRegistry;
 import org.springframework.web.servlet.LocaleResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;
 import org.springframework.web.servlet.i18n.SessionLocaleResolver;
-import ru.geekbrains.spring.ishop.rest.converters.*;
 
 import java.util.Locale;
 
@@ -42,16 +40,6 @@ public class AppConfig implements WebMvcConfigurer {
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(localeChangeInterceptor());
-    }
-
-    // TODO Delete
-    @Override
-    public void addFormatters(FormatterRegistry registry) {
-        registry.addConverter(new ProductToOutProductConverter());
-        registry.addConverter(new OrderToOutOrderConverter());
-        registry.addConverter(new OrderItemToOutOrderItemConverter());
-        registry.addConverter(new DeliveryToOutDeliveryConverter());
-        registry.addConverter(new UserToOutUserConverter());
     }
 
 }
